+++ src/RecoveryServices/RecoveryServices.Backup.Test/ScenarioTests/AzureWorkload/ContainerTests.ps1
@@ -28,7 +28,8 @@ function Get-AzureVmWorkloadContainer
          -ResourceId $resourceId `
          -BackupManagementType AzureWorkload `
          -WorkloadType MSSQL `
-         -VaultId $vault.ID
+         -VaultId $vault.ID `
+		 -Force
 	  Assert-AreEqual $container.Status "Registered"
 
       # VARIATION-1: Get All Containers with only mandatory parameters
@@ -81,13 +82,16 @@ function Unregister-AzureWorkloadContainer
          -ResourceId $resourceId `
          -BackupManagementType AzureWorkload `
          -WorkloadType MSSQL `
-         -VaultId $vault.ID
+         -VaultId $vault.ID `
+		 -Force
 	  Assert-AreEqual $container.Status "Registered"
 
 	  #Unregister container
-      Unregister-AzRecoveryServicesBackupContainer `
-		-VaultId $vault.ID `
-		-Container $container
+      Get-AzRecoveryServicesBackupContainer `
+         -VaultId $vault.ID `
+         -ContainerType AzureVMAppContainer `
+         -Status Registered `
+         -FriendlyName $containerName | Unregister-AzRecoveryServicesBackupContainer -VaultId $vault.ID
 
 	  $container = Get-AzRecoveryServicesBackupContainer `
          -VaultId $vault.ID `