+++ src/Monitor/Monitor/ChangeLog.md
@@ -26,6 +26,9 @@
       - Add-AzMetricAlertRuleV2
   * Updated Monitor SDK to version 0.22.0-preview
 
+## Version 1.0.1
+* Update help for Get-AzMetric
+
 ## Version 1.0.0
 * General availability of `Az.Monitor` module
 * Removed plural names "Categories" and "Timegrains" parameter in favor of singular parameter names